requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.45
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1

# Async Tasks
celery==5.3.4
kombu==5.3.4

# AWS Services
boto3==1.34.0
botocore==1.34.0

# Blockchain
web3==6.11.0
eth-account==0.8.0

# PDF & QR Generation
reportlab==4.0.4
qrcode[pil]==7.4.2
Pillow==10.4.0

# AI Integration
openai==1.3.0
httpx==0.25.2
stripe

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# Monitoring
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
python-dateutil==2.8.2
greenlet==3.0.1

 # Browser automation for screenshots
 playwright
